feat: implement user authentication and authorization in posts and users routes, enhance UI with dynamic username display

+++ routers/posts.py
@@ -9,6 +9,8 @@ import models
 from database import get_db
 from schemas import PostResponse, PostUpdate, PostCreate
 
+from auth import CurrentUser
+
 router = APIRouter()
 
 @router.get("", response_model=list[PostResponse])
@@ -20,17 +22,11 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
     return posts
 
 @router.post("", response_model=PostResponse, status_code=status.HTTP_201_CREATED)
-async def create_post(post: PostCreate, db: Annotated[AsyncSession,Depends(get_db)]):
-    result = await db.execute(select(models.User).where(models.User.id == post.user_id))
-    user = result.scalars().first()
-
-    if not user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
-
+async def create_post(post: PostCreate, current_user: CurrentUser, db: Annotated[AsyncSession,Depends(get_db)]):
     new_post = models.Post(
         title = post.title,
         content = post.content,
-        user_id = post.user_id
+        user_id = current_user.id
     )
 
     db.add(new_post)
@@ -47,35 +43,37 @@ async def get_post(post_id: int, db: Annotated[AsyncSession,Depends(get_db)]):
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Post with id {post_id} not found")
 
 @router.put("/{post_id}", response_model=PostResponse)
-async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[AsyncSession,Depends(get_db)]):
+async def update_post_full(post_id: int, post_data: PostCreate, current_user: CurrentUser, db: Annotated[AsyncSession,Depends(get_db)]):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Post with id {post_id} not found")
 
-    if post_data.user_id != post.user_id:
-        # if its not the same then check for whichever user they are trying to assign
-        # the post that even exixts or not.
-        result = await db.execute(select(models.User).where(models.User.id == post_data.user_id))
-        user = result.scalars().first()
-
-        if not user:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post"
+        )
 
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
+    post.user_id = current_user.id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
     return post
 
 @router.patch("/{post_id}", response_model=PostResponse)
-async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated[AsyncSession,Depends(get_db)]):
+async def update_post_partial(post_id: int, post_data: PostUpdate, current_user: CurrentUser, db: Annotated[AsyncSession,Depends(get_db)]):
     result = await db.execute(select(models.Post).options(selectinload(models.Post.author)).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Post with id {post_id} not found")
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post")
 
     update_data = post_data.model_dump(exclude_unset=True)
     for field, value in update_data.items():
@@ -87,11 +85,17 @@ async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated
 
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession,Depends(get_db)]):
+async def delete_post(post_id: int, current_user: CurrentUser, db: Annotated[AsyncSession,Depends(get_db)]):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Post with id {post_id} not found")
+    
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(
+        status_code=status.HTTP_403_FORBIDDEN,
+        detail="Not authorized to update this post")
 
     await db.delete(post)
     await db.commit()
